Move performance protocol definition to component
