+++ sdk/patina/src/uefi/protocol.rs
@@ -11,5 +11,3 @@
 //!
 //! SPDX-License-Identifier: Apache-2.0
 //!
-
-pub mod performance_measurement;
